refactor(desktop): modularize Rust backend, extract frontend hooks, and UX polish
Split lib.rs (1182 lines) into 10 focused modules: state, tray, commands
(system_info, sse, auth, window), and compositor (hyprland, sway, kwin).
Extract 4 hooks from __root.tsx (useTheme, useAuthState, useChannelActions,
useWidgetActions), reducing it from 784 to ~470 lines. Add tsc --noEmit to
build, fix Cargo.toml version mismatch, remove SSE URL hardcoding, centralize
channel/widget ordering, delete dead TickerSettings component, fix 14 empty
alt attributes, fix widget pin keyboard accessibility, and add errorComponent
to 5 routes.
