+++ desktop/src-tauri/src/commands/mod.rs
@@ -1,0 +1,4 @@
+pub mod auth;
+pub mod sse;
+pub mod system_info;
+pub mod window;
