+++ desktop/src/routes/settings.tsx
@@ -11,7 +11,20 @@ import { resetCategory } from "../preferences";
 
 export const Route = createFileRoute("/settings")({
   component: SettingsRoute,
+  errorComponent: SettingsError,
 });
+
+function SettingsError({ error }: { error: Error }) {
+  return (
+    <div className="flex flex-col items-center justify-center h-full text-center max-w-sm mx-auto gap-3 p-6">
+      <div className="w-10 h-10 rounded-xl bg-error/10 flex items-center justify-center mb-1">
+        <span className="text-error text-lg font-bold">!</span>
+      </div>
+      <h2 className="text-base font-semibold text-fg">Something went wrong</h2>
+      <p className="text-sm text-fg-3 leading-relaxed">{error.message}</p>
+    </div>
+  );
+}
 
 function SettingsRoute() {
   const shell = useShell();
